+++ examples/resources/incident_workflow/resource.tf
@@ -56,3 +56,55 @@ resource "incident_workflow" "autoassign_incident_lead" {
   ]
   state = "draft"
 }
+
+# This is a manually triggered workflow that collects information from the user
+# via form fields when they run it. The collected values are then available in
+# the workflow scope for use by conditions and steps.
+resource "incident_workflow" "page_execs" {
+  name    = "Page execs"
+  trigger = "manual"
+  expressions = [
+  ]
+  condition_groups = [
+  ]
+  steps = [
+    {
+      id   = "01HY0QG9WT62CEYJN8JD74MJNR" # This is the ID of the step in the workflow, and must be a ULID
+      name = "slack.send_message"
+      param_bindings = [
+        {
+          value = {
+            reference = "incident.slack_channel"
+          }
+        },
+        {
+          array_value = [
+            {
+              reference = "workflow_form.reason"
+            }
+          ]
+        },
+      ]
+    },
+  ]
+  form_fields = [
+    {
+      name        = "Reason for paging"
+      type        = "Text"
+      description = "Why are we paging the execs?"
+      placeholder = "e.g. Major customer-facing outage"
+      array       = false
+      required    = true
+    },
+  ]
+  once_for = [
+    "incident",
+  ]
+  include_private_incidents = false
+  continue_on_step_error    = false
+  runs_on_incidents         = "newly_created_and_active"
+  runs_on_incident_modes = [
+    "standard",
+  ]
+  state = "draft"
+}